+++ IL2-SR-Client/Singletons/RciStatus.cs
@@ -1,0 +1,11 @@
+namespace Ciribob.IL2.SimpleRadio.Standalone.Client.Singletons
+{
+    public enum RciStatus
+    {
+        None,
+        FriendlyOnly,
+        EnemyOnly,
+        Both,
+        Neutral
+    }
+}