+++ IL2-SR-Common/Network/UpdaterChecker.cs
@@ -22,7 +22,7 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Common
 
         public static readonly string MINIMUM_PROTOCOL_VERSION = "1.0.0.0";
 
-        public static readonly string VERSION = "1.0.3.6";
+        public static readonly string VERSION = "1.0.3.7";
 
         private static readonly Logger _logger = LogManager.GetCurrentClassLogger();